# specs/001-furniture-quotation-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 家具報價單系統 API
  description: |
    家具報價單系統後端 API，提供 PDF 上傳、BOQ 解析、Excel 報價單產出功能。

    ## 功能概述
    - PDF 檔案上傳與管理
    - BOQ 表格自動解析（使用 Gemini AI）
    - 平面圖數量核對
    - 惠而蒙格式 Excel 報價單產出

    ## 錯誤處理
    所有錯誤回應皆使用繁體中文訊息。
  version: 1.0.0
  contact:
    name: API 支援
servers:
  - url: http://localhost:8000
    description: 開發環境
  - url: http://api.example.com
    description: 生產環境
...
paths:
...
  /api/documents:
    post:
      tags:
        - Upload
      summary: 上傳 PDF 檔案
      description: |
        上傳單一或多個 PDF 檔案，系統將建立處理任務。

        **限制**：
        - 單檔最大 50MB
        - 最多同時上傳 5 個檔案
        - 僅接受 PDF 格式
      operationId: uploadFiles
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: PDF 檔案（最多 5 個）
                  maxItems: 5
                document_type:
                  type: string
                  enum: [boq, floor_plan, auto]
                  default: auto
                  description: 文件類型（auto 表示自動偵測）
              required:
                - files
      responses:
        '201':
          description: 上傳成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: 上傳失敗（檔案格式錯誤、大小超過限制等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 伺服器錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    get:
      tags:
        - Upload
      summary: 取得已上傳文件列表
      operationId: listDocuments
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: 回傳數量限制
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
          description: 篩選解析狀態
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceDocument'
                  total:
                    type: integer
  /api/documents/{document_id}:
    get:
      tags:
        - Upload
      summary: 取得單一文件資訊
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceDocument'
        '404':
          description: 找不到文件
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Upload
      summary: 刪除文件
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 刪除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 文件已刪除
        '404':
          description: 找不到文件
...
  /api/images/{image_id}:
    get:
      tags:
        - Upload
      summary: 取得提取的圖片
      operationId: getImage
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 圖片檔案
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: 找不到圖片
components:
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 文件 ID
...
  schemas:
...
    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
        documents:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
                format: uuid
              filename:
                type: string
              file_size:
                type: integer
              task_id:
                type: string
                format: uuid
                description: 自動解析任務 ID（若啟用）
        message:
          type: string
          example: 已成功上傳 3 個檔案
    SourceDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
        document_type:
          type: string
          enum: [boq, floor_plan, unknown]
        parse_status:
          type: string
          enum: [pending, processing, completed, failed]
        parse_progress:
          type: integer
          minimum: 0
          maximum: 100
        parse_message:
          type: string
        total_pages:
          type: integer
        extracted_items_count:
          type: integer
        extracted_images_count:
          type: integer
        uploaded_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        error_code:
          type: string
          example: FILE_001
        message:
          type: string
          example: 檔案大小超過限制（最大 50MB）
        detail:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time